sql, sem/tree: enhance the NodeFormatter interface
This patch introduces a new FmtCtx struct to replace the previous
passing of separate *bytes.Buffer and FmtFlags arguments.

This work was initially motivated by the need for a subsequent patch
to pass more data to some formatting functions without further
clobbering FmtFlags (which has grown unwieldly large already).

However in the process of doing so, some minor bugs were fixed (SCRUB
and WITH not anonymizing their identifiers properly) and the code was
generally polished.

Release note: None