sql,pgwire: rework the implementation of the Copy-in pgwire subprotocol
The COPY... FROM STDIN statement is special: it puts the connection into
a special mode allowing the client to keep sending data.
Before this patch, the implementation of this protocol was very
convoluted - the v3Conn, Executor, Session, Copy node and CopyData
statement all collaborated to get things going: the connection would
send the COPY FROM stmt to the Executor, which would execute it and
return a result which had the side-effect of shifting the conn to a state
machine that would produce (with the collaboration of the Session)
synthetic CopyData statements. These statements were executed and they'd
accumulate data in a session-bound object. Finally, the connection would
send a special instance of the CopyData statement that would clean the
session.

This intertwining of the connection and the execution layer is a problem
for the upcoming rewrite of these modules, which insists on a clean
separation between them.
This patch moves to an implementation that recognizes COPY FROM as a
special instruction. This instruction does not have a plan; instead, it
is supported by a struct that takes ownership of the pgwire connection
and encapsulates the buffering of data, execution of the insertNodes and
aoptionally management of transactions (if executed outside of an
explicit transaction) and network communication. Control of the
connection is handed back to the pgwire server when the subprotocol
finishes.

This patch also happens to fix two issues with the previous code:
1) something was wrong with the reporting of the command tag and the
ServerMsgCommandComplete msg. This could be seen by running a copy with
psql and noticing that nothing was being printed to the console after
the final "\.". postgres and the new code print "COPY <rows>".
2) the memory used for batching rows was accounted to the wrong monitor
as #21278 explains.

Release note: None