sql: teach the distSQLReceiver to discriminate between error types
Today our sql execution code doesn't discriminated between query
execution errors and errors in communication with the client (e.g. when
a query produces a result row and tries to send it to the client, at
which point it finds out that the client network connection is toast).
Communication errors are important because they mean that the sql
session is also toast - we should clean it up and stop processing
queries. So, communication errors, which originate in pgwire, then flow
all the way through sql execution back into pgwire, which recognizes
them and dismantles the session.
This scheme is already pretty funky today, and doesn't fit at all in the
future connExecutor code -> there a session is not directly driven by a
pgwire connection; the session runs its own goroutine communicating
async with pgwire. The connExecutor will be aware of the two different
types of errors and handle them differently.

In preparation for that, this patch makes code around the
distSQLReceiver able to discriminate between the two types of error - it
keeps track of communication errors versus all other errors.
The capability is not used yet in earnest, as both types of errors flow
to pgwire just like they did before, but the refactoring already
simplifies some code.

Release note: none